// src/app/modules/openapi/model/category-dto.ts
/**
 * E-Commerce Application
 *
 * Contact: sad@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductDTO } from './product-dto';


export interface CategoryDTO { 
    categoryId: number;
    categoryName: string;
    products?: Array<ProductDTO>;
}