// src/app/modules/openapi/model/new-product-dto.ts
/**
 * E-Commerce Application
 *
 * Contact: sad@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategoryNameDTO } from './category-name-dto';


export interface NewProductDTO { 
    productName: string;
    description: string;
    quantity?: number;
    price?: number;
    categories?: Array<CategoryNameDTO>;
}